bash_profile: Don't hardcode perl bin paths
I mean, it's fine, it's just me, but this is more portable, and maybe even helpful in ssh or down the line if I ever move from macports to perlbrew.

`perl -V` is pretty helpful, and grabbing `vendorbin` and `sitebin` is easy enough (removing the trailing space and wrapped quotes is a drag but fine).  It's not as clear what to do for man pages though: manpath likes the top directory AFAICT, while `perl -V` has `man1dir`, `vendorman1dir`, `siteman1dir` (and the same for man3).