update: Add debug option to show timestamps of functions
Throw all the `xargs` calls through a function to determine if we should show the timestamp or not, likewise the colorization stuff.  In the future, could add a calculation feature.  Also add -B to skip updating macports base and the sync tree.  Not enabled for everything, homebrew-cask is the biggest offender.