emacs: Turn of flymake_perlcritic
Old, creaky, meant relying on an older (and patched) version of flymake (https://github.com/flymake/emacs-flymake), which was dumb.  Should just scrap everything and use flycheck, maybe this will finally be the kick in the pants?  flymake-cursor had to go too, it was ancient and relying on the old/patched flymake from illusori (https://github.com/flymake/emacs-flymake-cursor), but I wasn't even really using it?

https://www.flycheck.org/en/28/_downloads/flycheck.html#Perl