emacs: `require` a few things, for the sake of simplicity
Also add another "google" to `webjump`, so it's listed first (or at least before google translate).  Has to be slightly different to replace the prior art, hence the leading `https`.

Would be nice to have a smarter way of sorting these.  They're ripe for some sort of list iteration, but that just seems a real pain to deal with alists, simple-query, and all that.