.perlcriticrc: Exclude duplicate Freenode critic config
So, `Perl::Critic::Community` includes `Perl::Critic::Freenode` and functions as an alias for it, the old name.  It's a good policy set, but ends up providing duplicate items to `perlcritic` and FlyCheck, which is a right pain.  Excluding those is easy enough, and there's no loss, but it's a silly thing to have to do IMO.