.bashrc: Set DELTA_FEATURES to enable cli usage of delta like icdiff
As noted, delta rocks.  I like the side-by-side view, so let's use it!  The actual config resides in (.local-).gitconfig, but of course the light/dark-mode process from d738752 won't work, so we once again duplicate (quadruplicate) it here.  This, however, won't reload automatically whenever the macOS Terminal profile changes (see brief discussion in 975ec48 and TIL), so unless/until I swap modes, this is where we are.  It's mitigated by not being a big deal and by being barely used.