.emacs: Remove auto-dark workaround, just use defaults for now
My dumb workaround from c727021 to allow me to just use the default behavior for the time being was dumb, and the default of leuven isn't so bad (except where it is, maybe that will encourage me to figure things out?).  At any rate, can also use standard-themes, they aren't so bad (especially the light) and do a fine job of replicating the basics.