.emacs: require a bunch of stuff, slowing down init
Basically reverts changes from a decade ago, but using emacsclient means I don't really care about startup time that much, so going from 3.2 to 4.1 seconds isn't a big deal.  I'm genuinely curious which are the biggest faults, but whatever.  The main point here is that this reduces a number of `assignment to free variable` warnings, which I care more about on an ongoing basis.

Bikeshedding.